入力プロンプトのテスト機能、おすすめ表示機能

+++ js/promptEdit.js
@@ -44,6 +44,39 @@ export const deleteColmun = async(table, type) => {
   }
 }
 
+/*
+document.getElementById("pass-button").addEventListener("click",async () =>{
+  const id = document.getElementById("id-input").value;
+  const pass = document.getElementById("pass-input").value;
+
+  let id_hash; 
+  let pass_hash;
+  await sha256(id).then((hash) => {
+    id_hash = hash;
+  });
+  await sha256(pass).then((hash)=> {
+    pass_hash = hash;
+  });
+
+  //確認処理(ハッシュ値は仮)
+  if(id_hash == "cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e" && pass_hash == "cf83e1357eefb8bdf1542850d66d8007d620e4050b5715dc83f4a921d36ce9ce47d0d13c5d85f2b0ff8318d2877eec2f63b931bd47417a81a538327af927da3e"){
+    document.getElementById("pass").style.display = "none"
+    document.getElementById("view").style.display = "block";
+
+    sessionStorage.setItem("id",id_hash);
+    sessionStorage.setItem("pass",pass_hash);
+  }
+});
+*/
+
+const sha256 = async function (text) {
+  const unit8 = new TextEncoder().encode(text);
+  const digest = await crypto.subtle.digest('SHA-512', unit8);
+
+  return Array.from(new Uint8Array(digest)).map(v => v.toString(16).padStart(2,"0")).join('');
+}
+
+const add_button = document.getElementById("add-button");
 document.addEventListener("DOMContentLoaded", async() => {
   const type_table = document.getElementById("type-table");
   const type_item = await getColmun("explanation", "simple_content");
@@ -65,6 +98,7 @@ document.addEventListener("DOMContentLoaded", async() => {
     const edit_td = document.createElement("td");
     const edit_button = document.createElement("input");
     edit_button.setAttribute("type","button");
+    edit_button.setAttribute("id","edit-button");
     edit_button.setAttribute("value", "📝");
     edit_button.addEventListener("click",() => {
       sessionStorage.setItem("editType",item.type);
@@ -77,15 +111,19 @@ document.addEventListener("DOMContentLoaded", async() => {
     const delete_td = document.createElement("td");
     const delete_button = document.createElement("input");
     delete_button.setAttribute("type","button");
+    delete_button.setAttribute("id","delete-button");
     delete_button.setAttribute("value","🗑️");
     delete_button.addEventListener("click",async () => {
       if(confirm(item.type + "を削除します。よろしいですか？")){
-        //explanation
-        await deleteColmun("explanation",item.type);
-        //prompts
-        await deleteColmun("prompts",item.type);
+        if(confirm("一度削除した項目は復元できません。" + item.type +"を削除してもよろしいですか？")){
+          //explanation
+          await deleteColmun("explanation",item.type);
+          //prompts
+          await deleteColmun("prompts",item.type);
 
-        alert(item.type + "を削除しました。");
+          alert(item.type + "を削除しました。");
+          location.reload();
+        }
       }
     });
     delete_td.appendChild(delete_button);
@@ -93,9 +131,10 @@ document.addEventListener("DOMContentLoaded", async() => {
 
     type_table.appendChild(tr);
   });
+  add_button.style.display = "inline";
 });
 
-document.getElementById("add-button").addEventListener("click", () => {
+add_button.addEventListener("click", () => {
   sessionStorage.removeItem("editType");
   window.open("promptChange.html","_blank");
 });
